Added WUnderground class
Pass in key, city and state. At startup and then on every hour it requests current time and today's
forecast from Weather Underground. It parses out the time and current, high and low temps then
calls callback() and sets a Ticker for the next update.

The feed() method needs to be called periodically (usually in loop()) to allow the weather to be
fetched in the foreground when the timer goes off because you can't access the network in a
timer callback.