Added ButtonManager class
Handles Debounce, Press/Release, Click and LongPress. All timing can be set in ctor. After
press is debounced a Press event is generated. If a button is released within ClickTime a
Click event is sent. If LongPressTime expired while pressed, a LongPress event is sent. If
the button is held down longer than ClickTime but shorter than LongPressTime a Release
event is sent.

ButtonManager can have multiple buttons with addButton. Each button has a unique ID which
can be used to identify it in the event callback. You can also set whether or not the INPUT_PULLUP
is enabled and whether the button is active high or low.