Got rid of passing Font and scrollString from Max7219Display
Now you just call showString, which takes a single string param. If the
string starts with the control character \a it uses the compact font rather
than the normal one. If it starts with \v it scrolls instead of shows. Control
characters can be combined in any order, but they must only appear at the
start of the string.

Scroll rates are fixed