Added "common" next state table
Now you can pass a vector of next states (the same as passed to addState())
which will be checked in all states first on a call to sendInput(). If found it will
goto the associated state. This gives you the ability to have a breakout function
that can happen no matter what state you're in