Added support for underscores to showString()
If the string is short enough to fit on the screen (no Watusi) and it doesn't scroll
then the characters in the string indicated by the underscoreStart and underscoreLength
params are underscored. An underscore is a line at the bottom of the screen under the
appropriate character, so it's really only practical when using the compact font, which leaves
space at the bottom of the screen.