Work in Progress - Added Application class to ESPLib
I'm splitting the general Application functionality out of the Clock class. This will allow me to use this library in applications that don't have a clock. This class will be subclassed by the main app. It will do network setup and will manage the state machine to allow network restart and reset. It can be passed an output pin having an LED that it will blink to show current network status. It can also be passed a button input pin that will manage a button that can be used to put the system in Restart or Reset mode.