Got Time server working
The time coming from the weather server is super inaccurate, so I switched back t o getting time from timezonedb. It's all still in TimeWeatherServer, but now it does separate calls. But now I get the timezone from the weather server from the zip code so you still only need the zip code to get the right time and weather. I had to switch from using localtime() to gmtime(). The time epoch I'm getting is actually the localtime. But on Arduino the timezone is not set so localtime and gmtime return the same values. But on the Mac it is set so I need to use gmtime which ignores the timezone.

This gets EtherClock working correctly.