Improved Brightness computations
Hysteresis was all wrong. It was taking the diff between the min and max brightness values, dividing that by Hysteresis and then only changing brightness if it changes more than that. what you want is to use Hysteresis as a percentage of the diff between min and max. I also increased the rate of reading the brightness to make it more accurate. Now brightness doesn't throb but it gives a value that is appropriate for every brightness level