Made WiFIManagerParameters into a vector and got rid of zipcode specific logic in Application
Now you add WiFIManagerParameters to a vector in Application and it finds the one you need and does all the management in that vector. That way there doesn't have to be application specific values in the generic code, like zipcode. Also added use of Format because it's easier to print values that way.