Do lazy startup of blink timer
Some devices (like apparently the C3) have a problem with global constructors. In this case the attach_ms() function is never called because it's in the Blinker ctor which is called from the global ctor of PostLightController. I could create PostLightController using new but that would put it in the heap. So instead I lazy start the timer the first time setRate is called