Got IDF to build
This also adds the wifimanager to the esp-idf build. This taught me that the Arduino build isn't that much more bloated than the esp-idf build. The real problem is that all the WiFi and WiFiManager stuff is very large. The Arduino version is only about 100K larger and that's without all the features of the Arduino WiFiManager. So my next step will be to push all this work to a branch and start over with an esp-idf build that includes the Arduino component so I can use all the Arduino features. We'll see what that comes up with