Got rid of a bunch of cross platform hackery
This is mostly switching from my String class, which wraps std::string on Mac, to simply using std::string everywhere. The ESP32 fully supports std::string and it simplifies a lot of stuff. I just had to add a couple of places where an Arduino API returns a String to use c_str() to convert to a raw string which then gets converted to a std::string. But this is all down in the Arduino specific code in ESPWiFiPortal.

I also got rid of ToString and ToFloat and converted a couple of Arduino named string functions to std variants. I also got rid of the F() macro because the ESP32 doesn't need it.

Everything works as before and is much cleaner