Rearranged how extra pages are included
Now Application looks for a file called system.html. If found then the top button on the front page says "System" and takes you to that page. Otherwise it links to the built-in filemgr page.

I rearranged endpoints a bit. Now the /fs endpoint is used to link to static pages in the / folder of LittleFS. So /fs/system.html opens /system.html. And now the file manager is at the /filemgr endpoint. A new /shell endpoint has been added but for now it's not implemented