Fixed some bugs in the LittleFSShim
good() is not the right value to use to test if the read or write failed. It is false when at eof. We want to use fail(), which is what operator bool tests, so we just use that.

Also return ints from read and write so we can return -1 for failure. And return  the correct length of the last read, which is stored in gcount().