Implemented more functionality for MacWebServer
I added an index.html file which is initially loaded in place of the WiFiManager portal page on ESP. It just has a link to the System page, which is the first button on the portal page on ESP. This index.html page comes up when you access the root of the web server (localhost). When you click on the button it takes you to the system.html page. Clicking on the filemgr or shell buttons still not working