Added move ctor/assignment and tightened up File class
This is what was causing all the problems. When assigning the File object after creation it did a copy which closed the file. Now it moves, which is more efficient and doesn't delete the underlying FILE object. I also deleted the copy ctor/assignment so they don't cause a prolbem.