Fixed bugs in uploading multiple files
This does a few things. First, when you upload the last file (or the only file) there is an extra '--' at the end of the boundary. Check for that and exit the loop when you see it. Second, when uploading multiple files I was adding a '\n--' to the boundary multiple times which was causing subsequent uploads to fail. Finally, I needed to distinguish between the last upload which has the extra -- and the others which don't